moving .h to .hpp and updating includes
