+++ include/rviz_visual_tools/tf_visual_tools.h
@@ -45,8 +45,8 @@
 #include <string>
 
 // ROS
-#include <ros/ros.h>
-#include <geometry_msgs/TransformStamped.h>
+#include <rclcpp/rclcpp.hpp>
+#include <geometry_msgs/msg/transform_stamped.hpp>
 
 // Eigen
 #include <Eigen/Geometry>
